# backend/docs/openapi/settlement-reward.json
{"openapi":"3.0.1","info":{"title":"PRE(D)ATA Backend API","description":"PRE(D)ATA V2 backend OpenAPI (service-grouped)","contact":{"name":"PRE(D)ATA Backend"},"version":"v2"},"servers":[{"url":"http://127.0.0.1:8080","description":"Local"}],"tags":[{"name":"settlement-reward","description":"Settlement and reward APIs"},{"name":"settlement-reward","description":"Premium data APIs"},{"name":"settlement-reward","description":"Analytics APIs"},{"name":"settlement-reward","description":"Reward APIs"},{"name":"settlement-reward","description":"Data analysis APIs"},{"name":"settlement-reward","description":"Reward admin APIs"}],"paths":{"/api/premium-data/preview":{"post":{"tags":["settlement-reward"],"operationId":"previewPremiumData","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PremiumDataRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopePremiumDataResponse"}}}}}}},"/api/premium-data/export":{"post":{"tags":["settlement-reward"],"operationId":"exportPremiumData","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PremiumDataRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopePremiumDataResponse"}}}}}}},"/api/analysis/questions/{id}/simulate-filter":{"post":{"tags":["settlement-reward"],"operationId":"simulateFilter","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FilterOptions"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopeFilterSimulationResponse"}}}}}}},"/api/admin/settlements/questions/{id}/settle":{"post":{"tags":["settlement-reward"],"operationId":"settleQuestion","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SettleQuestionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/admin/settlements/questions/{id}/settle-auto":{"post":{"tags":["settlement-reward"],"operationId":"settleQuestionAuto","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/admin/settlements/questions/{id}/finalize":{"post":{"tags":["settlement-reward"],"operationId":"finalizeSettlement","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/FinalizeSettlementRequest"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/admin/settlements/questions/{id}/cancel":{"post":{"tags":["settlement-reward"],"operationId":"cancelSettlement","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object"}}}}}}},"/api/admin/rewards/retry/{questionId}":{"post":{"tags":["settlement-reward"],"operationId":"retryFailedDistributions","parameters":[{"name":"questionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopeMapStringObject"}}}}}}},"/api/admin/rewards/distribute/{questionId}":{"post":{"tags":["settlement-reward"],"operationId":"distributeRewards","parameters":[{"name":"questionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopeMapStringObject"}}}}}}},"/api/settlements/history/me":{"get":{"tags":["settlement-reward"],"operationId":"getMySettlementHistory","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopeListSettlementHistoryItem"}}}}}}},"/api/rewards/{memberId}":{"get":{"tags":["settlement-reward"],"operationId":"getTotalRewards","parameters":[{"name":"memberId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopeTotalRewardResponse"}}}}}}},"/api/premium-data/quality-summary/{questionId}":{"get":{"tags":["settlement-reward"],"operationId":"getQualitySummary","parameters":[{"name":"questionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopeDataQualitySummary"}}}}}}},"/api/analytics/global/stats":{"get":{"tags":["settlement-reward"],"operationId":"getGlobalStats","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopeMapStringObject"}}}}}}},"/api/analytics/gap-analysis/{questionId}":{"get":{"tags":["settlement-reward"],"operationId":"getVoteBetGapAnalysis","parameters":[{"name":"questionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopeVoteBetGapReport"}}}}}}},"/api/analytics/filtering-effect/{questionId}":{"get":{"tags":["settlement-reward"],"operationId":"getFilteringEffect","parameters":[{"name":"questionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopeFilteringEffectReport"}}}}}}},"/api/analytics/demographics/{questionId}":{"get":{"tags":["settlement-reward"],"operationId":"getVoteDemographics","parameters":[{"name":"questionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopeVoteDemographicsReport"}}}}}}},"/api/analytics/dashboard/{questionId}":{"get":{"tags":["settlement-reward"],"operationId":"getQualityDashboard","parameters":[{"name":"questionId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopeQualityDashboard"}}}}}}},"/api/analysis/questions/{id}/weighted-votes":{"get":{"tags":["settlement-reward"],"operationId":"getWeightedVotes","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"category","in":"query","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopeWeightedVoteResult"}}}}}}},"/api/analysis/questions/{id}/quality-score":{"get":{"tags":["settlement-reward"],"operationId":"getQualityScore","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopeQualityScoreResponse"}}}}}}},"/api/analysis/questions/{id}/premium-data":{"get":{"tags":["settlement-reward"],"operationId":"getPremiumData","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"minLatencyMs","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":2000}},{"name":"onlyBettors","in":"query","required":false,"schema":{"type":"boolean","default":false}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopePremiumDataResponse"}}}}}}},"/api/analysis/questions/{id}/fast-clickers":{"get":{"tags":["settlement-reward"],"operationId":"getFastClickers","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"thresholdMs","in":"query","required":false,"schema":{"type":"integer","format":"int32","default":1000}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopeFastClickersResponse"}}}}}}},"/api/analysis/questions/{id}/by-country":{"get":{"tags":["settlement-reward"],"operationId":"getVotesByCountry","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopeObject"}}}}}}},"/api/analysis/questions/{id}/abusing-report":{"get":{"tags":["settlement-reward"],"operationId":"getAbusingReport","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiEnvelopeAbusingReport"}}}}}}}},"components":{"schemas":{"PremiumDataRequest":{"required":["questionId"],"type":"object","properties":{"questionId":{"type":"integer","format":"int64"},"countryCode":{"type":"string"},"jobCategory":{"type":"string"},"ageGroup":{"type":"integer","format":"int32"},"minTier":{"type":"string"},"minLatencyMs":{"type":"integer","format":"int32"}}},"ApiEnvelopePremiumDataResponse":{"required":["success","timestamp"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/PremiumDataResponse"},"error":{"$ref":"#/components/schemas/ErrorDetail"},"message":{"type":"string"},"timestamp":{"type":"string"}}},"ErrorDetail":{"required":["code","message","status"],"type":"object","properties":{"code":{"type":"string"},"message":{"type":"string"},"status":{"type":"integer","format":"int32"},"details":{"type":"array","items":{"type":"string"}}}},"PremiumDataPoint":{"required":["choice","countryCode","latencyMs","questionId","questionTitle","tier","tierWeight","timestamp","voteId"],"type":"object","properties":{"voteId":{"type":"integer","format":"int64"},"questionId":{"type":"integer","format":"int64"},"questionTitle":{"type":"string"},"choice":{"type":"string"},"latencyMs":{"type":"integer","format":"int32"},"countryCode":{"type":"string"},"jobCategory":{"type":"string"},"ageGroup":{"type":"integer","format":"int32"},"tier":{"type":"string"},"tierWeight":{"type":"number","format":"double"},"timestamp":{"type":"string"}}},"PremiumDataResponse":{"required":["data","filters","noCount","noPercentage","questionId","questionTitle","totalCount","yesCount","yesPercentage"],"type":"object","properties":{"questionId":{"type":"integer","format":"int64"},"questionTitle":{"type":"string"},"filters":{"$ref":"#/components/schemas/PremiumDataRequest"},"totalCount":{"type":"integer","format":"int32"},"yesCount":{"type":"integer","format":"int32"},"noCount":{"type":"integer","format":"int32"},"yesPercentage":{"type":"number","format":"double"},"noPercentage":{"type":"number","format":"double"},"data":{"type":"array","items":{"$ref":"#/components/schemas/PremiumDataPoint"}}}},"FilterOptions":{"required":["excludeCountries","minLatencyMs","minTierWeight","onlyBettors"],"type":"object","properties":{"minLatencyMs":{"type":"integer","format":"int32"},"onlyBettors":{"type":"boolean"},"minTierWeight":{"type":"number","format":"double"},"excludeCountries":{"type":"array","items":{"type":"string"}}}},"ApiEnvelopeFilterSimulationResponse":{"required":["success","timestamp"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/FilterSimulationResponse"},"error":{"$ref":"#/components/schemas/ErrorDetail"},"message":{"type":"string"},"timestamp":{"type":"string"}}},"FilterSimulationResponse":{"required":["filterOptions","filteredVoteCount","filteredYesPercentage","originalVoteCount","originalYesPercentage","percentageChange","questionId","removedCount"],"type":"object","properties":{"questionId":{"type":"integer","format":"int64"},"filterOptions":{"$ref":"#/components/schemas/FilterOptions"},"originalVoteCount":{"type":"integer","format":"int32"},"filteredVoteCount":{"type":"integer","format":"int32"},"removedCount":{"type":"integer","format":"int32"},"originalYesPercentage":{"type":"number","format":"double"},"filteredYesPercentage":{"type":"number","format":"double"},"percentageChange":{"type":"number","format":"double"}}},"SettleQuestionRequest":{"required":["finalResult"],"type":"object","properties":{"finalResult":{"type":"string"},"sourceUrl":{"type":"string"}}},"FinalizeSettlementRequest":{"required":["force"],"type":"object","properties":{"force":{"type":"boolean"}}},"ApiEnvelopeMapStringObject":{"required":["success","timestamp"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","additionalProperties":{"type":"object"}},"error":{"$ref":"#/components/schemas/ErrorDetail"},"message":{"type":"string"},"timestamp":{"type":"string"}}},"ApiEnvelopeListSettlementHistoryItem":{"required":["success","timestamp"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"array","items":{"$ref":"#/components/schemas/SettlementHistoryItem"}},"error":{"$ref":"#/components/schemas/ErrorDetail"},"message":{"type":"string"},"timestamp":{"type":"string"}}},"SettlementHistoryItem":{"required":["betAmount","finalResult","isWinner","myChoice","payout","profit","questionId","questionTitle"],"type":"object","properties":{"questionId":{"type":"integer","format":"int64"},"questionTitle":{"type":"string"},"myChoice":{"type":"string"},"finalResult":{"type":"string"},"betAmount":{"type":"integer","format":"int64"},"payout":{"type":"integer","format":"int64"},"profit":{"type":"integer","format":"int64"},"isWinner":{"type":"boolean"}}},"ApiEnvelopeTotalRewardResponse":{"required":["success","timestamp"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/TotalRewardResponse"},"error":{"$ref":"#/components/schemas/ErrorDetail"},"message":{"type":"string"},"timestamp":{"type":"string"}}},"TotalRewardResponse":{"required":["currentBalance","estimatedRewardPerVote","memberId","tier","tierWeight","totalVotes"],"type":"object","properties":{"memberId":{"type":"integer","format":"int64"},"currentBalance":{"type":"integer","format":"int64"},"totalVotes":{"type":"integer","format":"int32"},"tier":{"type":"string"},"tierWeight":{"type":"number","format":"double"},"estimatedRewardPerVote":{"type":"integer","format":"int64"}}},"ApiEnvelopeDataQualitySummary":{"required":["success","timestamp"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/DataQualitySummary"},"error":{"$ref":"#/components/schemas/ErrorDetail"},"message":{"type":"string"},"timestamp":{"type":"string"}}},"DataQualitySummary":{"required":["byLatency","byTier","questionId","totalVotes"],"type":"object","properties":{"questionId":{"type":"integer","format":"int64"},"totalVotes":{"type":"integer","format":"int32"},"byLatency":{"$ref":"#/components/schemas/LatencyBreakdown"},"byTier":{"type":"object","additionalProperties":{"type":"integer","format":"int32"}}}},"LatencyBreakdown":{"required":["fast","instant","normal","slow"],"type":"object","properties":{"instant":{"type":"integer","format":"int32"},"fast":{"type":"integer","format":"int32"},"normal":{"type":"integer","format":"int32"},"slow":{"type":"integer","format":"int32"}}},"ApiEnvelopeVoteBetGapReport":{"required":["success","timestamp"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/VoteBetGapReport"},"error":{"$ref":"#/components/schemas/ErrorDetail"},"message":{"type":"string"},"timestamp":{"type":"string"}}},"DistributionData":{"required":["noCount","noPercentage","yesCount","yesPercentage"],"type":"object","properties":{"yesCount":{"type":"integer","format":"int32"},"noCount":{"type":"integer","format":"int32"},"yesPercentage":{"type":"number","format":"double"},"noPercentage":{"type":"number","format":"double"}}},"VoteBetGapReport":{"required":["betDistribution","gapPercentage","qualityScore","questionId","voteDistribution"],"type":"object","properties":{"questionId":{"type":"integer","format":"int64"},"voteDistribution":{"$ref":"#/components/schemas/DistributionData"},"betDistribution":{"$ref":"#/components/schemas/DistributionData"},"gapPercentage":{"type":"number","format":"double"},"qualityScore":{"type":"number","format":"double"}}},"ApiEnvelopeFilteringEffectReport":{"required":["success","timestamp"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/FilteringEffectReport"},"error":{"$ref":"#/components/schemas/ErrorDetail"},"message":{"type":"string"},"timestamp":{"type":"string"}}},"FilteringData":{"required":["noPercentage","totalCount","yesPercentage"],"type":"object","properties":{"totalCount":{"type":"integer","format":"int32"},"yesPercentage":{"type":"number","format":"double"},"noPercentage":{"type":"number","format":"double"}}},"FilteringEffectReport":{"required":["afterFiltering","beforeFiltering","filteredCount","filteredPercentage","questionId"],"type":"object","properties":{"questionId":{"type":"integer","format":"int64"},"beforeFiltering":{"$ref":"#/components/schemas/FilteringData"},"afterFiltering":{"$ref":"#/components/schemas/FilteringData"},"filteredCount":{"type":"integer","format":"int32"},"filteredPercentage":{"type":"number","format":"double"}}},"AgeVoteData":{"required":["ageGroup","noCount","total","yesCount","yesPercentage"],"type":"object","properties":{"ageGroup":{"type":"integer","format":"int32"},"yesCount":{"type":"integer","format":"int32"},"noCount":{"type":"integer","format":"int32"},"total":{"type":"integer","format":"int32"},"yesPercentage":{"type":"number","format":"double"}}},"ApiEnvelopeVoteDemographicsReport":{"required":["success","timestamp"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/VoteDemographicsReport"},"error":{"$ref":"#/components/schemas/ErrorDetail"},"message":{"type":"string"},"timestamp":{"type":"string"}}},"CountryVoteData":{"required":["countryCode","noCount","total","yesCount","yesPercentage"],"type":"object","properties":{"countryCode":{"type":"string"},"yesCount":{"type":"integer","format":"int32"},"noCount":{"type":"integer","format":"int32"},"total":{"type":"integer","format":"int32"},"yesPercentage":{"type":"number","format":"double"}}},"JobVoteData":{"required":["jobCategory","noCount","total","yesCount","yesPercentage"],"type":"object","properties":{"jobCategory":{"type":"string"},"yesCount":{"type":"integer","format":"int32"},"noCount":{"type":"integer","format":"int32"},"total":{"type":"integer","format":"int32"},"yesPercentage":{"type":"number","format":"double"}}},"VoteDemographicsReport":{"required":["byAge","byCountry","byJob","questionId","totalVotes"],"type":"object","properties":{"questionId":{"type":"integer","format":"int64"},"totalVotes":{"type":"integer","format":"int32"},"byCountry":{"type":"array","items":{"$ref":"#/components/schemas/CountryVoteData"}},"byJob":{"type":"array","items":{"$ref":"#/components/schemas/JobVoteData"}},"byAge":{"type":"array","items":{"$ref":"#/components/schemas/AgeVoteData"}}}},"ApiEnvelopeQualityDashboard":{"required":["success","timestamp"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/QualityDashboard"},"error":{"$ref":"#/components/schemas/ErrorDetail"},"message":{"type":"string"},"timestamp":{"type":"string"}}},"QualityDashboard":{"required":["demographics","filteringEffect","gapAnalysis","overallQualityScore","questionId"],"type":"object","properties":{"questionId":{"type":"integer","format":"int64"},"demographics":{"$ref":"#/components/schemas/VoteDemographicsReport"},"gapAnalysis":{"$ref":"#/components/schemas/VoteBetGapReport"},"filteringEffect":{"$ref":"#/components/schemas/FilteringEffectReport"},"overallQualityScore":{"type":"number","format":"double"}}},"ApiEnvelopeWeightedVoteResult":{"required":["success","timestamp"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/WeightedVoteResult"},"error":{"$ref":"#/components/schemas/ErrorDetail"},"message":{"type":"string"},"timestamp":{"type":"string"}}},"WeightedVoteResult":{"required":["effectiveVotes","rawYesPercentage","totalVotes","weightedYesPercentage"],"type":"object","properties":{"rawYesPercentage":{"type":"number","format":"double"},"weightedYesPercentage":{"type":"number","format":"double"},"totalVotes":{"type":"integer","format":"int32"},"effectiveVotes":{"type":"number","format":"double"}}},"ApiEnvelopeQualityScoreResponse":{"required":["success","timestamp"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/QualityScoreResponse"},"error":{"$ref":"#/components/schemas/ErrorDetail"},"message":{"type":"string"},"timestamp":{"type":"string"}}},"QualityScoreResponse":{"required":["grade","qualityScore","questionId"],"type":"object","properties":{"questionId":{"type":"integer","format":"int64"},"qualityScore":{"type":"number","format":"double"},"grade":{"type":"string"}}},"ApiEnvelopeFastClickersResponse":{"required":["success","timestamp"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/FastClickersResponse"},"error":{"$ref":"#/components/schemas/ErrorDetail"},"message":{"type":"string"},"timestamp":{"type":"string"}}},"FastClickersResponse":{"required":["percentage","questionId","suspiciousCount","thresholdMs"],"type":"object","properties":{"questionId":{"type":"integer","format":"int64"},"thresholdMs":{"type":"integer","format":"int32"},"suspiciousCount":{"type":"integer","format":"int32"},"percentage":{"type":"number","format":"double"}}},"ApiEnvelopeObject":{"required":["success","timestamp"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object"},"error":{"$ref":"#/components/schemas/ErrorDetail"},"message":{"type":"string"},"timestamp":{"type":"string"}}},"AbusingReport":{"required":["overallGap","questionId","recommendation","suspiciousGroups","suspiciousMembers","totalMembers"],"type":"object","properties":{"questionId":{"type":"integer","format":"int64"},"suspiciousGroups":{"type":"array","items":{"$ref":"#/components/schemas/SuspiciousGroup"}},"overallGap":{"type":"number","format":"double"},"totalMembers":{"type":"integer","format":"int32"},"suspiciousMembers":{"type":"integer","format":"int32"},"recommendation":{"type":"string"}}},"ApiEnvelopeAbusingReport":{"required":["success","timestamp"],"type":"object","properties":{"success":{"type":"boolean"},"data":{"$ref":"#/components/schemas/AbusingReport"},"error":{"$ref":"#/components/schemas/ErrorDetail"},"message":{"type":"string"},"timestamp":{"type":"string"}}},"SuspiciousGroup":{"required":["betYesPercentage","criteria","gapIndex","memberCount","riskLevel","voteYesPercentage"],"type":"object","properties":{"criteria":{"type":"string"},"voteYesPercentage":{"type":"number","format":"double"},"betYesPercentage":{"type":"number","format":"double"},"gapIndex":{"type":"number","format":"double"},"memberCount":{"type":"integer","format":"int32"},"riskLevel":{"type":"string","enum":["LOW","MEDIUM","HIGH","CRITICAL"]}}}}}}